requirements.txt
altgraph==0.17.4
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
filelock==3.17.0
fsspec==2025.2.0
idna==3.10
Jinja2==3.1.5
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
packaging==24.2
pefile==2023.2.7
pydub==0.25.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pywin32-ctypes==0.2.3
regex==2024.11.6
requests==2.32.3
setuptools==75.8.2
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
